+++ CHANGELOG.md
@@ -5,6 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2023-09-28
+
+## Added
+
+- **Native**: Added a demo example.
+- **Java**: Added a demo example.
+- **Java**: Color activation choice option (support non-colored shell).
+
+## Changed
+
+- **Java**: Code rewriting and optimization.
+- **Java**: Full regex support for some options.
+- **Java**: Improved debug information.
+- **Java**: Readme update.
+
+## Fixed
+
+- **Java**: Fix support for all types of variables.
+- **Java**: Fix support for multi-constructor of a function.
+- **Java**: Fix method search.
+
 ## [1.2.0] - 2023-09-22
 
 ## Added
@@ -105,6 +126,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial Release.
 
+[1.2.1]: https://github.com/hyugogirubato/Frida-CodeShare/releases/tag/v1.2.1
 [1.2.0]: https://github.com/hyugogirubato/Frida-CodeShare/releases/tag/v1.2.0
 [1.1.5]: https://github.com/hyugogirubato/Frida-CodeShare/releases/tag/v1.1.5
 [1.1.4]: https://github.com/hyugogirubato/Frida-CodeShare/releases/tag/v1.1.4